Introduce `cairo-lang-doc` crate and move all item doc/sig/md code there
commit-id:793cceae
