+++ .github/CODEOWNERS
@@ -1,2 +1,3 @@
+/crates/cairo-lang-doc @mkaput @piotmag769
 /crates/cairo-lang-language-server @mkaput
 /vscode-cairo @mkaput
