+++ Cargo.toml
@@ -29,6 +29,7 @@ members = [
     "crates/cairo-lang-debug",
     "crates/cairo-lang-defs",
     "crates/cairo-lang-diagnostics",
+    "crates/cairo-lang-doc",
     "crates/cairo-lang-eq-solver",
     "crates/cairo-lang-filesystem",
     "crates/cairo-lang-formatter",
@@ -84,7 +85,7 @@ ark-secp256r1 = "0.4.0"
 ark-std = "0.4.0"
 assert_matches = "1.5"
 bimap = "0.6.3"
-cairo-vm = { version = "1.0.0-rc3", features = ["mod_builtin"]}
+cairo-vm = { version = "1.0.0-rc3", features = ["mod_builtin"] }
 clap = { version = "4.5.4", features = ["derive"] }
 colored = "2.1.0"
 const-fnv1a-hash = "1.1.0"
