+++ crates/cairo-lang-doc/Cargo.toml
@@ -1,0 +1,17 @@
+[package]
+name = "cairo-lang-doc"
+version.workspace = true
+edition.workspace = true
+repository.workspace = true
+license-file.workspace = true
+description = "A collection of documentation processing utilities for the Cairo programming language."
+
+[dependencies]
+cairo-lang-defs = { path = "../cairo-lang-defs", version = "~2.6.4" }
+cairo-lang-syntax = { path = "../cairo-lang-syntax", version = "~2.6.4" }
+cairo-lang-utils = { path = "../cairo-lang-utils", version = "~2.6.4" }
+salsa.workspace = true
+itertools.workspace = true
+
+[dev-dependencies]
+indoc.workspace = true
